Turn Nixpkgs options into consts
